ov006/dScMgHanachan_c: promote 22 shards into one link-verified TU
src/actors/dScMgHanachan_c.cpp absorbs 22 one-function sources covering
0x020eb018..0x020ec458 (ROM ordinals 13..34 of tu_map's 61-function run).

  tubuild verify   : 22/22 MATCH, objisolate clean, reloc-destinations clean,
                     emission order ROM-ascending -> TEXT-VERIFIED
  tubuild linkcheck: LINK-VERIFIED (pre-promotion). [4b/8] 22 LICENSED,
                     0 COLLIDES-GAP, 0 COLLIDES-SRC, 0 HOMELESS; module
                     fidelity 106/106 exact, 100.000000% of compared bytes;
                     ROM sha256 IDENTICAL to the stock control.
                     `dsd check symbols --fail` FAILs on the same 9
                     pre-existing errors the baseline control has: 0 new.

This is a deliberate partial, 22 of 61, for two measured reasons.

1. A SOURCELESS HOLE splits the run. ROM ordinal 11, func_ov006_020ea914
   (0x020ea914, size 0x324), has no source in the tree: it is a banked
   near-miss (config/match_attempts.jsonl, best 113 divergences) and has no
   entry in ov006/delinks.txt at all, so the cartridge's own bytes cover it.
   It sits in the MIDDLE of the run, and no delink block and no TU manifest
   anywhere in this repository carries two .text runs -- both checked
   tree-wide -- so a claim with a hole in it cannot be expressed. The run can
   only be licensed as one of its two contiguous sides.

2. TWO FILE-GLOBAL PRAGMAS CONTRADICT EACH OTHER on the larger side. Four of
   its 49 functions reproduce only with `#pragma opt_strength_reduction off`
   (020eac38, 020ec458, 020ec93c, 020ec9c0 -- each carried it in its own
   legacy file) and three only with it ON (020ebf20, 020ec2bc, 020ecba4).
   Render carried `#pragma opt_common_subs off`, which costs five other
   members when set. Both settings are file-global last-wins in mwccarm
   2004/b56, re-measured here: bracketing opt_strength_reduction off/on
   around exactly the four members that want it produced byte-for-byte the
   same 41/49 as omitting it, so there is no positional escape. Whole-side
   scores with every other reconciliation held identical:
     SR off / OCS on  44/49      SR on / OCS on  42/49      SR off / OCS off  40/49
   Promotion needs a CONTIGUOUS all-matching run; the longest over all three
   settings is ordinals 13..34 under SR on, which is what this file licenses.
   The other 27 functions keep their shards and their delinks.txt entries.

NOT a destructor-order refusal: the ROM orders this class D1 (0x020ea280)
below D0 (0x020ea2c8), the reproducible direction. The pair is out only
because it lies on the far side of the sourceless hole. Because it is out,
this TU does not own the class's key function (the first declared virtual is
the out-of-line ~dScMgHanachan_c()), so it emits no _ZTV/_ZTI/_ZTS, needs no
compiler_only_output block -- verify reports zero unlicensed symbols -- and
needed no header edit.

Reconciliation notes worth keeping:
  - Every 2-int shadow struct that is ever whole-struct-assigned is spelled
    with an ARRAY member. C++ scalarizes an assignment between structs of
    named ints word-by-word where C block-moved it; that alone accounted for
    seven of the DIFFs (Pair, P, P2, W2, Words2, and two Vec2i copies).
  - ApproachLinear must be declared OUTSIDE the extern "C" block. Inside it
    the mangling is stripped and the module link fails `Undefined:
    "ApproachLinear"` -- verify stays 22/22 either way, so only linkcheck
    catches it.
  - Three members' parameter types were changed to agree with
    include/decl_common.h (020eac38 void*, 020eb558 char*, 020ed8a4 void*);
    a disagreeing spelling is `illegal function overloading` reported against
    the definition with no mention of the header.

layout_check clean, check_dead_references no new, check_tubuild_conflicts
clean (all 22 notes marked RESOLVED and their markers deleted),
cpp_tu_compat --require-ready READY, eligible 11095/11234 with every member
of this TU eligible, tiers_ratchet --update wrote no change and no backslide
row (none of the 22 auto-named shards was in the CONVERTED set).
