--- src/func_ov006_020eb9b0.c
@@ -1,9 +1,0 @@
-void func_ov006_020eb9b0(char *p)
-{
-    *(int *)(p + 0x44) = 2560;
-    *(short *)(p + 0x86) = 1536;
-    *(short *)(p + 0x8e) = 0;
-    *(char *)(p + 0x93) = 0;
-    *(char *)(p + 0x94) = 0;
-    *(char *)(p + 0x95) = 2;
-}
